Tidy up travis_test.sh and ensure tests actually run
Currently, the tests aren't running on Travis, as we "cd snappy" and
never "cd ..", meaning that py.test discovery doesn't find any tests to
run.

This commit tidies up the `travis_test.sh` script a little, improves the
output on Travis by folding non-test-related output, and ensures that
we're in the correct directory when we run py.test.