Support Python 2.6 struct without compromising on unicode_literals
Using `from __future__ import unicode_literals` ensures that the use of
bytestrings vs. text is consistent between Python 2.x and Python 3.x --
it means, for example, that we don't accidentally add code for Python
2.x which deals in str objects, which suddenly become unicode in Python
3.

Unfortunately, the `struct` package in Python 2.6 poses a problem, as it
only accepts `str` values for its `fmt` argument, and raises a TypeError
if you pass it `unicode`. Using

  b'bytestring literals'

is an option, but it seems ugly when the parameter is clearly text (even
if it is guaranteed to be ASCII). It seems wrong to make this concession
for Python 2.6 when Python 2.7 and 3.x do the right thing.

So, we introduce a compat layer for `struct.pack` and `struct.unpack`
which can be removed once support for Python 2.6 is dropped.